admin and login related templates refactoring
