+++ views/admin/addPostView.php
@@ -90,4 +90,4 @@
   
 <?php $content = ob_get_clean(); ?>
 
-<?php require_once 'template.php'; ?>
+<?php require_once 'admin_template.php'; ?>
