+++ views/admin/editPostView.php
@@ -91,4 +91,4 @@
 
 <?php $content = ob_get_clean(); ?>
 
-<?php require_once 'template.php'; ?>
+<?php require_once 'admin_template.php'; ?>
