+++ views/admin/homeAdminView.php
@@ -165,4 +165,4 @@
 
 <?php $content = ob_get_clean(); ?>
 
-<?php require_once 'template.php'; ?>
+<?php require_once 'admin_template.php'; ?>
