+++ views/admin/loginView.php
@@ -1,28 +1,10 @@
+<?php $title = "Login" ;?>
 
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title>Mon Blog | Login</title>
-  <!-- Tell the browser to be responsive to screen width -->
-  <meta name="viewport" content="width=device-width, initial-scale=1">
+<?php ob_start(); ?>
 
-  <!-- Font Awesome -->
-  <link rel="stylesheet" href="<?=URL_PATH;?>public/admin/plugins/fontawesome-free/css/all.min.css">
-  <!-- Ionicons -->
-  <link rel="stylesheet" href="https://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css">
-  <!-- icheck bootstrap -->
-  <link rel="stylesheet" href="<?=URL_PATH;?>public/admin/plugins/icheck-bootstrap/icheck-bootstrap.min.css">
-  <!-- Theme style -->
-  <link rel="stylesheet" href="<?=URL_PATH;?>public/admin/dist/css/adminlte.min.css">
-  <!-- Google Font: Source Sans Pro -->
-  <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700" rel="stylesheet">
-</head>
-<body class="hold-transition login-page">
 <div class="login-box">
   <div class="login-logo">
-    <a href="../../index2.html"><b>Log</b>in</a>
+    <a href="#"><b>Log</b>in</a>
   </div>
   <!-- /.login-logo -->
   <div class="card">
@@ -95,12 +77,6 @@
 </div>
 <!-- /.login-box -->
 
-<!-- jQuery -->
-<script src="<?=URL_PATH;?>public/admin/plugins/jquery/jquery.min.js"></script>
-<!-- Bootstrap 4 -->
-<script src="<?=URL_PATH;?>public/admin/plugins/bootstrap/js/bootstrap.bundle.min.js"></script>
-<!-- AdminLTE App -->
-<script src="<?=URL_PATH;?>public/admin/dist/js/adminlte.min.js"></script>
+<?php $content = ob_get_clean(); ?>
 
-</body>
-</html>
+<?php require_once 'login_template.php'; ?>
