+++ views/admin/userView.php
@@ -1,42 +1,6 @@
-<?php
+<?php $title = "Utilisateurs inscrits"; ?>
 
-  if(!isset($_SESSION['user_id']))
-  {
-    header('Location: '. URL_PATH.'authentication/login');    
-      exit();
-  }
-?>
-
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title><?php echo WEBSITENAME; ?> | admins</title>
-  <!-- Tell the browser to be responsive to screen width -->
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-
-  <!-- Font Awesome -->
-  <link rel="stylesheet" href="<?php echo URL_PATH; ?>public/admin/plugins/fontawesome-free/css/all.min.css">
-  <!-- Ionicons -->
-  <link rel="stylesheet" href="https://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css">
-  <!-- DataTables -->
-  <link rel="stylesheet" href="<?php echo URL_PATH; ?>public/admin/plugins/datatables-bs4/css/dataTables.bootstrap4.min.css">
-  <link rel="stylesheet" href="<?php echo URL_PATH; ?>public/admin/plugins/datatables-responsive/css/responsive.bootstrap4.min.css">
-  <!-- Theme style -->
-  <link rel="stylesheet" href="<?php echo URL_PATH; ?>public/admin/dist/css/adminlte.min.css">
-  <!-- Google Font: Source Sans Pro -->
-  <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700" rel="stylesheet">
-</head>
-<body class="hold-transition sidebar-mini">
-<div class="wrapper">
-
-  <!-- Navbar -->
- <?php require_once('themeparts/topmenuAdmin.php'); ?>
-  <!-- /.navbar -->
-
-  <!-- Main Sidebar Container -->
-  <?php require_once('themeparts/sidebarAdmin.php'); ?>
+<?php ob_start(); ?>
 
   <!-- Content Wrapper. Contains page content -->
   <div class="content-wrapper">
@@ -124,47 +88,7 @@
     <!-- /.content -->
   </div>
   <!-- /.content-wrapper -->
-  
-  <?php require_once('themeparts/footerAdmin.php'); ?>
 
-  <!-- Control Sidebar -->
-  <aside class="control-sidebar control-sidebar-dark">
-    <!-- Control sidebar content goes here -->
-  </aside>
-  <!-- /.control-sidebar -->
-</div>
-<!-- ./wrapper -->
+<?php $content = ob_get_clean(); ?>
 
-<!-- jQuery -->
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/jquery/jquery.min.js"></script>
-<!-- Bootstrap 4 -->
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/bootstrap/js/bootstrap.bundle.min.js"></script>
-<!-- DataTables -->
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/datatables/jquery.dataTables.min.js"></script>
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/datatables-bs4/js/dataTables.bootstrap4.min.js"></script>
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/datatables-responsive/js/dataTables.responsive.min.js"></script>
-<script src="<?php echo URL_PATH; ?>public/admin/plugins/datatables-responsive/js/responsive.bootstrap4.min.js"></script>
-<!-- AdminLTE App -->
-<script src="<?php echo URL_PATH; ?>public/admin/dist/js/adminlte.min.js"></script>
-<!-- AdminLTE for demo purposes -->
-<script src="<?php echo URL_PATH; ?>public/admin/dist/js/demo.js"></script>
-<!-- page script -->
-<script>
-  $(function () {
-    $("#example1").DataTable({
-      "responsive": true,
-      "autoWidth": false,
-    });
-    $('#example2').DataTable({
-      "paging": true,
-      "lengthChange": false,
-      "searching": false,
-      "ordering": true,
-      "info": true,
-      "autoWidth": false,
-      "responsive": true,
-    });
-  });
-</script>
-</body>
-</html>
+<?php require_once 'admin_template.php'; ?>
